fix chrome not being tracked and add decimals for quotas

+++ app/src/main/java/com/appblocker/BlockSet.kt
@@ -6,7 +6,7 @@ data class BlockSet(
     val id: String = UUID.randomUUID().toString(),
     var name: String = "",
     var apps: MutableList<String> = mutableListOf(), // package names
-    var quotaMinutes: Int = 30,
+    var quotaMinutes: Double = 30.0,
     var windowMinutes: Int = 60, // clock-aligned window: 5, 10, 15, 20, 30, or 60 minutes
     var combinedQuota: Boolean = true // if true, all apps share the quota
 )
